requirements.txt
psycopg2
django==2.2.5
djangorestframework==3.10
djoser
djangorestframework_simplejwt
pylint
pylint-django
python-social-auth==0.3.6
social-auth-app-django==3.1.0
social-auth-core==3.2.0
django-cors-headers
Pillow
django-imagekit
qrcode==6.1
django-countries
django_filters
django-extra-fields
